requirements.txt
tensorflow-gpu==2.0.0
pandas==0.25.0
sklearn==0.0
pillow==6.1.0
matplotlib==3.1.1
seaborn==0.9.0
requests==2.22.0